+++ workout_diary/src/entities/user.py
@@ -1,3 +1,5 @@
+import bcrypt
+
 class User:
     """Class, creates a new User.
 
@@ -10,3 +12,7 @@ class User:
     def __init__(self, username, password):
         self.username = username
         self.password = password
+
+
+    def __str__(self):
+        return f"Username: {self.username}, password (hashed): {self.password}"