+++ workout_diary/src/ui/new_workout.py
@@ -1,0 +1,3 @@
+class CreateNewWorkout:
+    def __init__(self) -> None:
+        pass